backend/src/ml/python/requirements.txt
# PowerStream ML Microservice Dependencies

# Web framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# ML/AI
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# API utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Monitoring
prometheus-flask-exporter>=0.22.0













